+++ pubspec.yaml
@@ -9,14 +9,17 @@ environment:
   sdk: ">=2.15.1 <3.0.0"
 
 dependencies:
+  archive: ^3.2.0
   bitsdojo_window: ^0.1.1+1
+  ffi: ^1.1.2
   fluent_ui: ^3.9.0
   flutter:
     sdk: flutter
   flutter_acrylic: ^1.0.0+1
+  path: ^1.8.0
   path_provider: ^2.0.9
-  tar: ^0.5.2
   very_good_analysis: ^2.4.0
+  win32: ^2.4.1
 
 dev_dependencies:
   flutter_test:
@@ -26,7 +29,6 @@ dev_dependencies:
 # The following section is specific to Flutter.
 flutter:
   uses-material-design: true
-
 # To add assets to your application, add an assets section, like this:
 # assets:
 #   - images/a_dot_burr.jpeg
